Split mod data checker.
