+++ game_ff12.py
@@ -19,13 +19,7 @@ from PyQt6.QtCore import (
 
 from PyQt6.QtWidgets import QMainWindow
 
-from ..basic_features import (
-    BasicLocalSavegames,
-    BasicModDataChecker,
-    GlobPatterns,
-)
-from ..basic_features.utils import is_directory
-
+from ..basic_features import BasicLocalSavegames
 from ..basic_features.basic_save_game_info import (
     BasicGameSaveGame,
     BasicGameSaveGameInfo,
@@ -38,81 +32,8 @@ from ..steam_utils import find_steam_path
 import vdf
 
 from .ff12.AutoUpdate import UpdateChecker
+from .ff12.ModDataChecker import FF12ModDataChecker
 from .ff12.SettingsManager import SettingsManager, settings_manager, SettingName
-
-class FF12ModDataChecker(BasicModDataChecker):
-    def __init__(self, organizer: mobase.IOrganizer, plugin_name: str):
-        self._organizer = organizer
-        self._plugin_name = plugin_name
-
-        super().__init__(
-            GlobPatterns(
-                unfold=['*'],
-                delete=["*"],
-                valid=["x64", "mods", "dxgi.dll", "dinput8.dll", "launcher.dll"],
-                move={"scripts":        "x64/",
-                      "modules":        "x64/",
-                      "gamedata":       "mods/deploy/ff12data/",
-                      "jsondata":       "mods/deploy/ff12data/",
-                      "prefetchdata":   "mods/deploy/ff12data/",
-                      "ps2data":        "mods/deploy/ff12data/",
-                      "ff12data":       "mods/deploy/",
-                      },
-            )
-        )
-
-    def dataLooksValid(
-        self, filetree: mobase.IFileTree
-    ) -> mobase.ModDataChecker.CheckReturn:
-        status = mobase.ModDataChecker.VALID
-
-        rp = self._regex_patterns
-        for entry in filetree:
-            name = entry.name().casefold()
-
-            if rp.valid.match(name):
-                if status is mobase.ModDataChecker.INVALID:
-                    status = mobase.ModDataChecker.VALID
-
-            elif rp.move_match(name) is not None:
-                status = mobase.ModDataChecker.FIXABLE
-
-            elif rp.unfold.match(name) and is_directory(entry):
-                status = mobase.ModDataChecker.FIXABLE
-                new_status = self.dataLooksValid(entry)
-                if new_status is not mobase.ModDataChecker.VALID:
-                    status = new_status
-
-            elif rp.delete.match(name) is not None:
-                status = mobase.ModDataChecker.FIXABLE
-
-            else:
-                status = mobase.ModDataChecker.INVALID
-                break
-        return status
-
-    def fix(self, filetree: mobase.IFileTree) -> mobase.IFileTree:
-        rp = self._regex_patterns
-
-        for entry in list(filetree):
-            name = entry.name().casefold()
-
-            if rp.valid.match(name):
-                continue
-
-            elif (move_key := rp.move_match(name)) is not None:
-                target = self._file_patterns.move[move_key]
-                filetree.move(entry, target)
-
-            elif rp.unfold.match(name) and is_directory(entry):
-                filetree.merge(entry)
-                entry.detach()
-                self.fix(filetree)
-
-            elif rp.delete.match(name):
-                entry.detach()
-
-        return filetree
 
 class FF12SaveGame(BasicGameSaveGame):
     def __init__(self, filepath: Path):
